etl/requirements.txt
backoff==2.2.1
certifi==2022.12.7
elastic-transport==8.4.0
elasticsearch==8.5.3
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
urllib3==1.26.13